More known issues. Small changes to make setup scripts more robust.

+++ setup-fedhost.sh
@@ -9,6 +9,8 @@ if [[ $# != 2 ]]; then
     exit 1
 fi  
 
+kubectl config use-context ${1}
+
 binding=$(kubectl get clusterrolebinding permissive-binding)
 if [[ -z $binding ]]; then
   echo creating permissive binding. NOT recommended for production clusters
@@ -28,16 +30,17 @@ fedName=myfederation
 
 set -e
 
-curl -LO https://dl.k8s.io/v1.7.0-alpha.4/kubernetes-client-linux-amd64.tar.gz
-tar xzf kubernetes-client-linux-amd64.tar.gz
-
-sudo cp kubernetes/client/bin/kubefed /usr/local/bin
-sudo chmod +x /usr/local/bin/kubefed
-sudo cp kubernetes/client/bin/kubectl /usr/local/bin
-sudo chmod +x /usr/local/bin/kubectl
+if [ ! -f kubernetes-client-linux-amd64.tar.gz ]; then
+  curl -LO https://dl.k8s.io/v1.7.0-alpha.4/kubernetes-client-linux-amd64.tar.gz
+  tar xzf kubernetes-client-linux-amd64.tar.gz
+  sudo cp kubernetes/client/bin/kubefed /usr/local/bin
+  sudo chmod +x /usr/local/bin/kubefed
+  sudo cp kubernetes/client/bin/kubectl /usr/local/bin
+  sudo chmod +x /usr/local/bin/kubectl
+fi
 
 wd=$(pwd)
-kubefed init ${fedName} --dns-provider="azure-azuredns" --dns-zone-name=${2} --dns-provider-config=${wd}/dns.conf --image=xtoph/hyperkube-amd64:azuredns.20
+kubefed init ${fedName} --dns-provider="azure-azuredns" --dns-zone-name=${2} --dns-provider-config=${wd}/dns.conf --image=xtoph/hyperkube-amd64:azuredns.20 --controllermanager-arg-overrides="--v=5"
 
 # double check
 ns=$(kubectl get namespace --context=myfederation)
@@ -46,12 +49,11 @@ if [[ -n $binding ]]; then
 fi
 
 kubectl config use-context ${fedName}
-
-
 kubefed join ${cluster} --host-cluster-context=${cluster} --cluster-context=${cluster}
 
-
-git clone https://github.com/OguzPastirmaci/hostname
+if [[ ! -d ./hostname ]]; then 
+  git clone https://github.com/OguzPastirmaci/hostname
+fi 
 
 kubectl get clusters
